Move Simulation Output after each Simulation
The previous version of this script ran superfish for all
of the supplied values and moved the output files at the end.
If the program ended prematurely the output files were not
moved to the appropriate directory. The output files are now
moved after each individual simulation.

I have also added an output to the terminal to indicate how many
simulations are left to complete.